e7.17 navigation menu (added it previously so just moved logout to it)
